Move FlexTripEdge to separate package
