+++ src/ext/java/org/opentripplanner/ext/flex/template/FlexAccessTemplate.java
@@ -1,6 +1,6 @@
 package org.opentripplanner.ext.flex.template;
 
-import org.opentripplanner.ext.flex.FlexTripEdge;
+import org.opentripplanner.ext.flex.edgetype.FlexTripEdge;
 import org.opentripplanner.ext.flex.flexpathcalculator.FlexPathCalculator;
 import org.opentripplanner.ext.flex.trip.FlexTrip;
 import org.opentripplanner.model.SimpleTransfer;
